fix -Wshorten-64-to-32 warnings
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@38153 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/c64/cart/ocean.c
@@ -95,7 +95,7 @@
 static uint8_t currbank = 0;
 static uint8_t regval = 0;
 static uint8_t io1_mask = 0x3f;
-static unsigned int cart_size = 0;
+static uint32_t cart_size = 0;
 
 /* ---------------------------------------------------------------------*/
 static void ocean_io1_store(uint16_t addr, uint8_t value)
@@ -198,13 +198,13 @@ static int ocean_common_attach(void)
 /* ---------------------------------------------------------------------*/
 
 /* HACK: 32k isnt really a valid size, see above */
-int ocean_cart_sizes[] = { 0x80000, 0x40000, 0x20000, 0x08000, 0 };
+uint32_t ocean_cart_sizes[] = { 0x80000, 0x40000, 0x20000, 0x08000, 0 };
 
 int ocean_bin_attach(const char *filename, uint8_t *rawcart)
 {
     int rc = -1;
     int i;
-    size_t size;
+    uint32_t size;
     for (i = 0; (size = ocean_cart_sizes[i]) != 0; i++) {
         rc = util_file_load(filename, rawcart, size, UTIL_FILE_LOAD_SKIP_ADDRESS);
         if (rc == 0) {
@@ -223,7 +223,7 @@ int ocean_bin_attach(const char *filename, uint8_t *rawcart)
 
 int ocean_crt_attach(FILE *fd, uint8_t *rawcart)
 {
-    size_t rom_size;
+    uint32_t rom_size;
     crt_chip_header_t chip;
 
     rom_size = 0;
